After live push

+++ site/app/Http/Controllers/Admin/ExerciseController.php
@@ -22,6 +22,7 @@ use App\Fundumental;
 use FFMpeg\FFMpeg;
 use FFMpeg\Coordinate\TimeCode as TimeCode;
 use App\Video;
+use Yajra\Datatables\Datatables;
 
 class ExerciseController extends Controller
 {
@@ -39,13 +40,59 @@ class ExerciseController extends Controller
      */
     public function getIndex()
     {
-        // Grab all the users
-        $exercise = Exercise::all();
+
 
         $user = User::where('id', Auth::user()->id)->with(['profile', 'settings'])->first();
 
         // Show the page
-        return View('admin.exercise.index', compact('exercise', 'user'));
+        return View('admin.exercise.index', compact('user'));
+    }
+
+    /**
+     * Process datatables ajax request.
+     *
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function anyData()
+    {
+        $posts = Exercise::select('exercises.*');
+        $db = DB::table('muscle_groups');
+        return Datatables::of($posts)
+                ->addColumn('action', function ($list) {
+                    $html = '<a href="' . route('admin.exercise.show', $list->id) . '"><i class="glyphicon glyphicon-eye-open" data-name="info" data-size="18" data-loop="true" data-c="#428BCA" data-hc="#428BCA" title="View Exercise Details"></i></a>
+                                <a href="' . route('admin.exercise.edit', $list->id) . '"><i class="glyphicon glyphicon-edit" data-name="edit" data-size="18" data-loop="true" data-c="#428BCA" data-hc="#428BCA" title="Edit Exercise Details"></i></a>
+
+                                <a href="' . route('admin.confirm-delete.exercise', $list->id) . '" data-toggle="modal" data-target="#delete_confirm">
+                                    <i class="glyphicon glyphicon-remove" data-name="exercise-remove" data-size="18" data-loop="true" data-c="#f56954" data-hc="#f56954" title="delete exercise">
+                                    </i>
+                                </a>';
+                    return $html;
+                })
+                ->editColumn('category', function ($list) {
+                    if ($list->category == 1) {
+                        return 'Lean';
+                    } elseif ($list->category == 2) {
+                        return 'Athletic';
+                    } else {
+                        return 'Strength';
+                    }
+                })
+                ->editColumn('type', function ($list) {
+                    if ($list->type == 1) {
+                        return 'Free';
+                    } else {
+                        return 'Paid';
+                    }
+                })
+                ->editColumn('unit', function ($list) {
+                    if ($list->unit == 'times') {
+                        return 'Repetitions';
+                    } else {
+                        return 'Seconds';
+                    }
+                })
+                ->blacklist(['action'])
+                ->make(true); 
     }
 
     /**
@@ -83,6 +130,100 @@ class ExerciseController extends Controller
             }
         }
 
+        if ((Input::get('pro_tips') != 'Type the pro tips here') && Input::get('pro_tips') != '') {
+            $proTipsRaw = str_replace(['<p>', '<div>'], '', Input::get('pro_tips'));
+            $proTipsRaw = str_replace(['</p>', '<br>', '<br />', '</div>'], '', $proTipsRaw);
+            preg_match_all('!<ul>(.+?)</ul>!s', $proTipsRaw, $ulmatches);
+
+            foreach ($ulmatches[1] AS $ustring) {
+                $ulStringRaw = str_replace(['<ul>', '</li>'], '', $ustring);
+                $ulStringRaw = str_replace('<li>', '', $ulStringRaw);
+                $proTipsRaw = str_replace($ulmatches[0], $ulStringRaw, $proTipsRaw);
+            }
+
+            preg_match_all('!<ol>(.+?)</ol>!s', $proTipsRaw, $olmatches);
+
+            foreach ($olmatches[1] AS $ostring) {
+                $olStringRaw = str_replace(['<ol>', '</li>'], '', $ostring);
+                $olStringRawArray = explode('<li>', $olStringRaw);
+                unset($olStringRawArray[0]);
+                for ($i = 1; $i <= count($olStringRawArray); $i++) {
+                    $olStringRawArray[$i] = $i . '. ' . $olStringRawArray[$i];
+                }
+                $olStringRaw = implode('', $olStringRawArray);
+                $proTipsRaw = str_replace($olmatches[0], $olStringRaw, $proTipsRaw);
+            }
+            $proTips = htmlspecialchars_decode(html_entity_decode($proTipsRaw), ENT_QUOTES);
+            $proTipsHtml = Input::get('pro_tips');
+            unset($proTipsRaw);
+        } else {
+            $proTips = '';
+            $proTipsHtml = '';
+        }
+
+        if ((Input::get('range_of_motion') != 'Type the range of motion here') && Input::get('range_of_motion') != '') {
+
+            $proTipsRaw = str_replace(['<p>', '<div>'], '', Input::get('range_of_motion'));
+            $proTipsRaw = str_replace(['</p>', '<br>', '<br />', '</div>'], '', $proTipsRaw);
+            preg_match_all('!<ul>(.+?)</ul>!s', $proTipsRaw, $ulmatches);
+
+            foreach ($ulmatches[1] AS $ustring) {
+                $ulStringRaw = str_replace(['<ul>', '</li>'], '', $ustring);
+                $ulStringRaw = str_replace('<li>', '', $ulStringRaw);
+                $proTipsRaw = str_replace($ulmatches[0], $ulStringRaw, $proTipsRaw);
+            }
+
+            preg_match_all('!<ol>(.+?)</ol>!s', $proTipsRaw, $olmatches);
+
+            foreach ($olmatches[1] AS $ostring) {
+                $olStringRaw = str_replace(['<ol>', '</li>'], '', $ostring);
+                $olStringRawArray = explode('<li>', $olStringRaw);
+                unset($olStringRawArray[0]);
+                for ($i = 1; $i <= count($olStringRawArray); $i++) {
+                    $olStringRawArray[$i] = $i . '. ' . $olStringRawArray[$i];
+                }
+                $olStringRaw = implode('', $olStringRawArray);
+                $proTipsRaw = str_replace($olmatches[0], $olStringRaw, $proTipsRaw);
+            }
+            $rangeOfMotion = htmlspecialchars_decode(html_entity_decode($proTipsRaw), ENT_QUOTES);
+            $rangeOfMotionHtml = Input::get('range_of_motion');
+            unset($proTipsRaw);
+        } else {
+            $rangeOfMotion = '';
+            $rangeOfMotionHtml = '';
+        }
+
+        if ((Input::get('video_tips') != 'Type the video tips here') && Input::get('video_tips') != '') {
+            $proTipsRaw = str_replace(['<p>', '<p></p>', '<p> </p>', '<p>&nbsp;</p>', '<div>'], '', Input::get('video_tips'));
+            $proTipsRaw = str_replace(['</p>', '<br>', '<br />', '</div>'], '', $proTipsRaw);
+            preg_match_all('!<ul>(.+?)</ul>!s', $proTipsRaw, $ulmatches);
+
+            foreach ($ulmatches[1] AS $ustring) {
+                $ulStringRaw = str_replace(['<ul>', '</li>'], '', $ustring);
+                $ulStringRaw = str_replace('<li>', '', $ulStringRaw);
+                $proTipsRaw = str_replace($ulmatches[0], $ulStringRaw, $proTipsRaw);
+            }
+
+            preg_match_all('!<ol>(.+?)</ol>!s', $proTipsRaw, $olmatches);
+
+            foreach ($olmatches[1] AS $ostring) {
+                $olStringRaw = str_replace(['<ol>', '</li>'], '', $ostring);
+                $olStringRawArray = explode('<li>', $olStringRaw);
+                unset($olStringRawArray[0]);
+                for ($i = 1; $i <= count($olStringRawArray); $i++) {
+                    $olStringRawArray[$i] = $i . '. ' . $olStringRawArray[$i];
+                }
+                $olStringRaw = implode('', $olStringRawArray);
+                $proTipsRaw = str_replace($olmatches[0], $olStringRaw, $proTipsRaw);
+            }
+            $videoTips = htmlspecialchars_decode(html_entity_decode($proTipsRaw), ENT_QUOTES);
+            $videoTipsHtml = Input::get('range_of_motion');
+            unset($proTipsRaw);
+        } else {
+            $videoTips = '';
+            $videoTipsHtml = '';
+        }
+
         $muscleGroups = Input::get('muscle_groups');
         $duration = Input::get('repetitions');
         $isStatic = Input::get('is_static');
@@ -91,16 +232,17 @@ class ExerciseController extends Controller
                 'description' => Input::get('description'),
                 'category' => Input::get('category'),
                 'type' => Input::get('type'),
-                'rewards' => Input::get('rewards'),
-                'repititions' => Input::get('repetitions'),
-                'duration' => $duration,
+                'rewards' => Input::get('rewards'),                
                 'unit' => Input::get('unit'),
                 'equipment' => Input::get('equipment'),
                 'muscle_groups' => (isset($muscleGroups) && !empty($muscleGroups != '')) ? implode(',', Input::get('muscle_groups')) : '',
-                'range_of_motion' => ((Input::get('range_of_motion') == 'Type the range of motion here') || Input::get('range_of_motion') == '') ? '' : Input::get('range_of_motion'),
-                'video_tips' => ((Input::get('video_tips') == 'Type the video tips here') || Input::get('video_tips') == '') ? '' : Input::get('video_tips'),
-                'pro_tips' => ((Input::get('pro_tips') == 'Type the pro tips here') || Input::get('pro_tips') == '') ? '' : Input::get('pro_tips'),
-                'is_static'  => (isset($isStatic)) ? 1 : 0
+                'range_of_motion' => $rangeOfMotion,
+                'video_tips' => $videoTips,
+                'pro_tips' => $proTips,
+                'range_of_motion_html' => str_replace("\r\n", '', $rangeOfMotionHtml),
+                'video_tips_html' => str_replace("\r\n", '', $videoTipsHtml),
+                'pro_tips_html' => str_replace("\r\n", '', $proTipsHtml),
+                'is_static' => (isset($isStatic)) ? 1 : 0
         ]);
 
         if (!is_null($exercise)) {
@@ -219,10 +361,104 @@ class ExerciseController extends Controller
         $exercise = Exercise::where('id', $id)->first();
 
         $muscleGroups = Input::get('muscle_groups');
-        
+
         $isStatic = Input::get('is_static');
 
         if (!is_null($exercise)) {
+
+            if ((Input::get('pro_tips') != 'Type the pro tips here') && Input::get('pro_tips') != '') {
+                $proTipsRaw = str_replace(['<p>', '<div>'], '', Input::get('pro_tips'));
+                $proTipsRaw = str_replace(['</p>', '<br>', '<br />', '</div>'], '', $proTipsRaw);
+                preg_match_all('!<ul>(.+?)</ul>!s', $proTipsRaw, $ulmatches);
+
+                foreach ($ulmatches[1] AS $ustring) {
+                    $ulStringRaw = str_replace(['<ul>', '</li>'], '', $ustring);
+                    $ulStringRaw = str_replace('<li>', '', $ulStringRaw);
+                    $proTipsRaw = str_replace($ulmatches[0], $ulStringRaw, $proTipsRaw);
+                }
+
+                preg_match_all('!<ol>(.+?)</ol>!s', $proTipsRaw, $olmatches);
+
+                foreach ($olmatches[1] AS $ostring) {
+                    $olStringRaw = str_replace(['<ol>', '</li>'], '', $ostring);
+                    $olStringRawArray = explode('<li>', $olStringRaw);
+                    unset($olStringRawArray[0]);
+                    for ($i = 1; $i <= count($olStringRawArray); $i++) {
+                        $olStringRawArray[$i] = $i . '. ' . $olStringRawArray[$i];
+                    }
+                    $olStringRaw = implode('', $olStringRawArray);
+                    $proTipsRaw = str_replace($olmatches[0], $olStringRaw, $proTipsRaw);
+                }
+                $proTips = htmlspecialchars_decode(html_entity_decode($proTipsRaw), ENT_QUOTES);
+                $proTipsHtml = Input::get('pro_tips');
+                unset($proTipsRaw);
+            } else {
+                $proTips = '';
+                $proTipsHtml = '';
+            }
+
+            if ((Input::get('range_of_motion') != 'Type the range of motion here') && Input::get('range_of_motion') != '') {
+
+                $proTipsRaw = str_replace(['<p>', '<div>'], '', Input::get('range_of_motion'));
+                $proTipsRaw = str_replace(['</p>', '<br>', '<br />', '</div>'], '', $proTipsRaw);
+                preg_match_all('!<ul>(.+?)</ul>!s', $proTipsRaw, $ulmatches);
+
+                foreach ($ulmatches[1] AS $ustring) {
+                    $ulStringRaw = str_replace(['<ul>', '</li>'], '', $ustring);
+                    $ulStringRaw = str_replace('<li>', '', $ulStringRaw);
+                    $proTipsRaw = str_replace($ulmatches[0], $ulStringRaw, $proTipsRaw);
+                }
+
+                preg_match_all('!<ol>(.+?)</ol>!s', $proTipsRaw, $olmatches);
+
+                foreach ($olmatches[1] AS $ostring) {
+                    $olStringRaw = str_replace(['<ol>', '</li>'], '', $ostring);
+                    $olStringRawArray = explode('<li>', $olStringRaw);
+                    unset($olStringRawArray[0]);
+                    for ($i = 1; $i <= count($olStringRawArray); $i++) {
+                        $olStringRawArray[$i] = $i . '. ' . $olStringRawArray[$i];
+                    }
+                    $olStringRaw = implode('', $olStringRawArray);
+                    $proTipsRaw = str_replace($olmatches[0], $olStringRaw, $proTipsRaw);
+                }
+                $rangeOfMotion = htmlspecialchars_decode(html_entity_decode($proTipsRaw), ENT_QUOTES);
+                $rangeOfMotionHtml = Input::get('range_of_motion');
+                unset($proTipsRaw);
+            } else {
+                $rangeOfMotion = '';
+                $rangeOfMotionHtml = '';
+            }
+
+            if ((Input::get('video_tips') != 'Type the video tips here') && Input::get('video_tips') != '') {
+                $proTipsRaw = str_replace(['<p>', '<p></p>', '<p> </p>', '<p>&nbsp;</p>', '<div>'], '', Input::get('video_tips'));
+                $proTipsRaw = str_replace(['</p>', '<br>', '<br />', '</div>'], '', $proTipsRaw);
+                preg_match_all('!<ul>(.+?)</ul>!s', $proTipsRaw, $ulmatches);
+
+                foreach ($ulmatches[1] AS $ustring) {
+                    $ulStringRaw = str_replace(['<ul>', '</li>'], '', $ustring);
+                    $ulStringRaw = str_replace('<li>', '', $ulStringRaw);
+                    $proTipsRaw = str_replace($ulmatches[0], $ulStringRaw, $proTipsRaw);
+                }
+
+                preg_match_all('!<ol>(.+?)</ol>!s', $proTipsRaw, $olmatches);
+
+                foreach ($olmatches[1] AS $ostring) {
+                    $olStringRaw = str_replace(['<ol>', '</li>'], '', $ostring);
+                    $olStringRawArray = explode('<li>', $olStringRaw);
+                    unset($olStringRawArray[0]);
+                    for ($i = 1; $i <= count($olStringRawArray); $i++) {
+                        $olStringRawArray[$i] = $i . '. ' . $olStringRawArray[$i];
+                    }
+                    $olStringRaw = implode('', $olStringRawArray);
+                    $proTipsRaw = str_replace($olmatches[0], $olStringRaw, $proTipsRaw);
+                }
+                $videoTips = htmlspecialchars_decode(html_entity_decode($proTipsRaw), ENT_QUOTES);
+                $videoTipsHtml = Input::get('range_of_motion');
+                unset($proTipsRaw);
+            } else {
+                $videoTips = '';
+                $videoTipsHtml = '';
+            }
 
             $duration = Input::get('repetitions');
             Exercise::where('id', $id)->update([
@@ -230,20 +466,21 @@ class ExerciseController extends Controller
                 'description' => Input::get('description'),
                 'category' => Input::get('category'),
                 'type' => Input::get('type'),
-                'rewards' => Input::get('rewards'),
-                'repititions' => Input::get('repetitions'),
-                'duration' => $duration,
+                'rewards' => Input::get('rewards'),                
                 'unit' => Input::get('unit'),
                 'equipment' => Input::get('equipment'),
                 'muscle_groups' => (isset($muscleGroups) && !empty($muscleGroups != '')) ? implode(',', Input::get('muscle_groups')) : '',
-                'range_of_motion' => ((Input::get('range_of_motion') == 'Type the range of motion here') || Input::get('range_of_motion') == '') ? '' : Input::get('range_of_motion'),
-                'video_tips' => ((Input::get('video_tips') == 'Type the video tips here') || Input::get('video_tips') == '') ? '' : Input::get('video_tips'),
-                'pro_tips' => ((Input::get('pro_tips') == 'Type the pro tips here') || Input::get('pro_tips') == '') ? '' : Input::get('pro_tips'),
-                'is_static'  => (isset($isStatic)) ? 1 : 0
+                'range_of_motion' => $rangeOfMotion,
+                'video_tips' => $videoTips,
+                'pro_tips' => $proTips,
+                'range_of_motion_html' => str_replace("\r\n", '', $rangeOfMotionHtml),
+                'video_tips_html' => str_replace("\r\n", '', $videoTipsHtml),
+                'pro_tips_html' => str_replace("\r\n", '', $proTipsHtml),
+                'is_static' => (isset($isStatic)) ? 1 : 0
             ]);
-            
+
             Fundumental::where('exercise_id', $id)->update(['unit' => Input::get('unit')]);
-            
+
             Workoutexercise::where('exercise_id', $id)->update(['unit' => Input::get('unit')]);
 
             if (isset($_FILES['video']) && $_FILES['video']['error'] == UPLOAD_ERR_OK) {
@@ -260,10 +497,14 @@ class ExerciseController extends Controller
                 if (!is_null($video)) {
 
                     //Unlink previusly uploaded file and thumbnail
-
-                    unlink(public_path('uploads/videos/') . $video->path);
-
-                    unlink(config("image.videoThumbPath") . $video->videothumbnail);
+                    
+                    if(file_exists(public_path('uploads/videos/') . $video->path)){
+                        unlink(public_path('uploads/videos/') . $video->path);
+                    }
+                    
+                    if(file_exists(config("image.videoThumbPath") . $video->videothumbnail)){
+                        unlink(config("image.videoThumbPath") . $video->videothumbnail);
+                    }
 
                     Video::where('parent_id', $exercise->id)->where('parent_type', 1)->update([
                         'user_id' => Auth::user()->id,
@@ -284,11 +525,11 @@ class ExerciseController extends Controller
                 }
             } else {
                 $video = Video::where('parent_id', $exercise->id)->where('parent_type', 1)->update([
-                        'description' => $exercise->name
+                    'description' => $exercise->name
                 ]);
             }
 
-            return Redirect::route('admin.exercises')->with('success', 'Updated successfully');
+            return Redirect::route('admin.exercises')->with('success', 'Updated successfully.');
         } else {
 
             return Redirect::route('admin.exercises', $id)->withInput()->with('error', 'Exercise not found!');
